Use appropriate data types for different memory accesses

+++ src/cdrom/core.hpp
@@ -72,11 +72,11 @@ class core_t : public memory_component_t {
  public:
   core_t(interrupt_access_t *irq, const char *game_file_name, bool log_enabled);
 
-  uint32_t io_read_byte(uint32_t address);
+  uint8_t io_read_byte(uint32_t address);
 
   uint32_t io_read_word(uint32_t address);
 
-  void io_write_byte(uint32_t address, uint32_t data);
+  void io_write_byte(uint32_t address, uint8_t data);
 
   void io_write_port_0_n(uint8_t data);
