+++ src/input/core.hpp
@@ -50,15 +50,15 @@ class core_t : public memory_component_t {
 
   void tick(int amount);
 
-  uint32_t io_read_byte(uint32_t address);
+  uint8_t io_read_byte(uint32_t address);
 
-  uint32_t io_read_half(uint32_t address);
+  uint16_t io_read_half(uint32_t address);
 
   uint32_t io_read_word(uint32_t address);
 
-  void io_write_byte(uint32_t address, uint32_t data);
+  void io_write_byte(uint32_t address, uint8_t data);
 
-  void io_write_half(uint32_t address, uint32_t data);
+  void io_write_half(uint32_t address, uint16_t data);
 
   void io_write_word(uint32_t address, uint32_t data);
