+++ src/memory-access.hpp
@@ -9,12 +9,12 @@ namespace psx {
 
 class memory_access_t {
  public:
-  virtual uint32_t read_byte(uint32_t address) = 0;
-  virtual uint32_t read_half(uint32_t address) = 0;
+  virtual uint8_t read_byte(uint32_t address) = 0;
+  virtual uint16_t read_half(uint32_t address) = 0;
   virtual uint32_t read_word(uint32_t address) = 0;
 
-  virtual void write_byte(uint32_t address, uint32_t data) = 0;
-  virtual void write_half(uint32_t address, uint32_t data) = 0;
+  virtual void write_byte(uint32_t address, uint8_t data) = 0;
+  virtual void write_half(uint32_t address, uint16_t data) = 0;
   virtual void write_word(uint32_t address, uint32_t data) = 0;
 };
