+++ src/memory-component.hpp
@@ -17,12 +17,12 @@ class memory_component_t {
 
   void log(const char *fmt, ...);
 
-  virtual uint32_t io_read_byte(uint32_t address);
-  virtual uint32_t io_read_half(uint32_t address);
+  virtual uint8_t io_read_byte(uint32_t address);
+  virtual uint16_t io_read_half(uint32_t address);
   virtual uint32_t io_read_word(uint32_t address);
 
-  virtual void io_write_byte(uint32_t address, uint32_t data);
-  virtual void io_write_half(uint32_t address, uint32_t data);
+  virtual void io_write_byte(uint32_t address, uint8_t data);
+  virtual void io_write_half(uint32_t address, uint16_t data);
   virtual void io_write_word(uint32_t address, uint32_t data);
 };
